backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# 配置管理
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI API 客户端
openai==1.3.7

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 邮箱验证
email-validator==2.1.0
